extract shellwords monkey patch
