+++ fastlane_core/lib/fastlane_core.rb
@@ -1,6 +1,7 @@
 require_relative 'fastlane_core/globals'
 # Ruby monkey-patches - should be before almost all else
 require_relative 'fastlane_core/core_ext/string'
+require_relative 'fastlane_core/core_ext/shellwords'
 
 require_relative 'fastlane_core/env'
 require_relative 'fastlane_core/feature/feature'
